get callback service refactor
