+++ cms-adapter/src/Rsbc.Dmf.CaseManagement/Manager/CaseManager.cs
@@ -2,7 +2,6 @@
 using Microsoft.Extensions.Logging;
 using Microsoft.OData.Client;
 using Rsbc.Dmf.CaseManagement.Dynamics;
-using Rsbc.Dmf.CaseManagement.Utilities;
 using Rsbc.Dmf.Dynamics.Microsoft.Dynamics.CRM;
 using Serilog;
 using System;
@@ -344,7 +343,6 @@ namespace Rsbc.Dmf.CaseManagement
             _caseMapper = caseMapper;
             OutcomeStatusTypes = GetOutcomeStatusTypes();
             OutcomeSubStatusTypes = GetOutcomeSubStatusTypes();
-
         }
 
         private Dictionary<string, Guid> GetOutcomeStatusTypes()
@@ -714,42 +712,6 @@ namespace Rsbc.Dmf.CaseManagement
             }
 
             return result;
-        }
-
-        ///<summary>
-        ///Get Driver Callbacks
-        /// </summary>
-        /// <param name="driverId"></param>
-        /// <returns></returns>
-        public async Task<IEnumerable<Callback>> GetDriverCallbacks(Guid driverId)
-        {
-            // get cases and include callbacks
-            var cases = dynamicsContext.incidents
-                .Expand(c => c.Incident_Tasks)
-                .Where(c => c._dfp_driverid_value == driverId && c.statecode == 0);
-
-            // compile a list of callbacks from each case
-            var results = new List<Callback>();
-            foreach (var @case in cases) 
-            {
-                // skip if tasks is null or has no active task
-                if (!@case.Incident_Tasks?.Any(task => task.statecode == 0) ?? false)
-                    break;
-
-                foreach (var callback in @case.Incident_Tasks.Where(task => task.statecode == 0)) 
-                {
-                    results.Add(new Callback
-                    {
-                        Id = callback.activityid ?? Guid.NewGuid(),
-                        RequestCallback = callback.scheduledend.GetValueOrDefault(),
-                        Topic = CallbackTopic.Upload,//Enum.Parse<CallbackTopic>(callback.activitytypecode), e.g. "task"
-                        CallStatus = (CallbackCallStatus)callback.statecode,
-                        Closed = callback.actualend.GetValueOrDefault()
-                    });
-                }
-            }
-
-            return results;
         }
 
         /// <summary>
